+++ RELEASE_NOTES.md
@@ -7,6 +7,8 @@
 
 **⚠ PLEASE UNDERSTAND:** Features and changes below may not fully work as the source code is constantly under development. For a stable and more tested version, please use the published binary releases.
 
+* Fixes a string/bool type error in server.js
+
 
 # Published releases